refactor(realm): extract hid bridge from js-realm-shared
